refactor(Character): 스타일을 별도의 파일로 분리
