+++ src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import { ROUTES } from './constants/routes';
 
 // pages
-import { default as CharacterIndex } from '@/pages/Character/Index';
+import { default as CharacterIndex } from '@/pages/Character/Character/Character';
 import CharacterLayout from '@/pages/Character/Layout';
 import DiariesDetail from '@/pages/Diaries/Detail';
 import DiariesFeedback from '@/pages/Diaries/Feedback';
